refactor: queryDSL 추가 적용

+++ src/main/java/com/elice/boardgame/game/dto/GameResponseDto.java
@@ -50,5 +50,5 @@ public class GameResponseDto {
 
     private List<GameGenre> gameGenres = new ArrayList<>();
 
-    private int views;
+    private Long views;
 }
